# [EVAKA] CI: refactor workflow for speed and parallelization