--- a/frontend/src/e2e-test/specs/2_employee-2/employee-dailynote.spec.ts
+++ b/frontend/src/e2e-test/specs/2_employee-2/employee-dailynote.spec.ts
@@ -44,7 +44,7 @@ let careArea: CareAreaBuilder
 
 const employeeHome = new EmployeeHome()
 
-fixture('Mobile daily notes')
+fixture('Mobile employee daily notes')
   .meta({ type: 'regression', subType: 'mobile' })
   .beforeEach(async () => {
     await resetDatabase()