--- a/frontend/src/e2e-test/specs/5_employee/income.spec.ts
+++ b/frontend/src/e2e-test/specs/5_employee/income.spec.ts
@@ -15,7 +15,7 @@ const fridgeHeadInformation = new FridgeHeadInformationPage()
 
 let personId: string
 
-fixture('Invoicing - invoices')
+fixture('Invoicing - income')
   .meta({ type: 'regression', subType: 'invoices' })
   .beforeEach(async (t) => {
     await resetDatabase()