--- a/message-service/build.gradle.kts
+++ b/message-service/build.gradle.kts
@@ -192,6 +192,9 @@ tasks {
     test {
         useJUnitPlatform()
         systemProperty("spring.profiles.active", "test")
+        filter {
+            setFailOnNoMatchingTests(false)
+        }
     }
 
     create("integrationTest", Test::class) {