--- a/service-lib/build.gradle.kts
+++ b/service-lib/build.gradle.kts
@@ -52,5 +52,8 @@ tasks.withType<KotlinCompile> {
 tasks {
     test {
         useJUnitPlatform()
+        filter {
+            setFailOnNoMatchingTests(false)
+        }
     }
 }