--- a/service/build.gradle.kts
+++ b/service/build.gradle.kts
@@ -200,7 +200,7 @@ tasks {
     }
 
     create("bootRunTest", org.springframework.boot.gradle.tasks.run.BootRun::class) {
-        main = "fi.espoo.evaka.MainKt"
+        mainClass.set("fi.espoo.evaka.MainKt")
         classpath = sourceSets["main"].runtimeClasspath
         systemProperty("spring.profiles.active", "local")
         systemProperty("spring.datasource.url", "jdbc:postgresql://localhost:15432/evaka_it")