@mikkopiu 2021-06-30
Verified that new tests (and tests that don't have timings) are randomly placed: the output is just a bit confusing:

```
No timing found for "fi.espoo.evaka.pis.AbstractIntegrationTest"
```

comes from the splitting going through _all_ the tests -> the same log will appear on all nodes **and one of the nodes will actually  receive those classes** ([example](https://app.circleci.com/pipelines/github/espoon-voltti/evaka/8032/workflows/581e916c-111d-40a6-b930-15921ff4ac06/jobs/148202/parallel-runs/3?filterBy=ALL&invite=true#step-105-32)). And just for further proof, here's a run with a totally new test class: https://app.circleci.com/pipelines/github/espoon-voltti/evaka/8032/workflows/581e916c-111d-40a6-b930-15921ff4ac06/jobs/148202/parallel-runs/0?filterBy=ALL&invite=true#step-105-30