@mikkopiu 2021-06-30
Improved test splitting output for debugging purposes, split Playwright tests (I think I just forgot about this after I started off the path of running E2E tests in a container...) and updated to the latest `voltti/builder-aws` images.